Correct a mismatch in the web three test modules, where we actually get an executable called ethrpctest, where we expect to get one called testweb3.   It looks like there is an ugly mixture of different semantics for how the test folders are used here, such that "ethrpctest" could be built standalone or as part of all the tests.
What we actually get on disk is ethrpctest, so that's what I am going to set the output executable name up as.    This change does not, I believe, modify any of the results of the build process, but it does remove the inconsistency in the build files, which has the nice side-effect of correcting the dependency graph, which is how I spotted this inconsistency in the first place.

Maybe as a follow-up we can consider actually changing the module name in test/CMakeLists.txt, so that you always get "testweb3".  If we do that then we should probably change the ethrpctest into a normal source folder at the same time.